few updates

+++ stock/IStockItem.cs
@@ -2,6 +2,6 @@ namespace el_chupitos_pattern.stock
 {
     interface IStockItem {
         string Name { get; set;}
-        int Price { get; set;}
+        double Price { get; set;}
     } 
 }
