+++ stock/confiture.cs
@@ -3,9 +3,9 @@ namespace el_chupitos_pattern.stock
     class Confiture : IStockItem
     {
         public string Name {get;set;}
-        public int Price {get;set;}
+        public double Price {get;set;}
 
-        public Confiture(string name, int price){
+        public Confiture(string name, double price){
             Name = name;
             Price = price;
         }
